Improvement for node and dependencies deletion action and DB relations

+++ src/NodeTypes/SectionNode.php
@@ -85,14 +85,15 @@ class SectionNode extends FolderNode
      */
     public function deleteNode() : bool
     {
+        // Deletes dependencies in rel tables
+        $depsMngr = new DepsManager();
+        if ($depsMngr->deleteBySource(DepsManager::SECTION_XIMLET, $this->parent->get('IdNode')) === false) {
+            return false;
+        }
         $section = new Section($this->parent->get('IdNode'));
         if ($section->delete() === false) {
             return false;
         }
-        
-        // Deletes dependencies in rel tables
-        $depsMngr = new DepsManager();
-        $depsMngr->deleteBySource(DepsManager::SECTION_XIMLET, $this->parent->get('IdNode'));
         Logger::info('Section dependencies deleted');
         return true;
     }
